pyproject.toml
[tool.poetry]
name = "figi-backend"
version = "0.0.0"
description = "Backend for Figi face search."
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
opencv-python = "^4.10.0.84"
fastapi = {extras = ["standard"], version = "^0.115.3"}
pgvector = "^0.3.5"
peewee = {extras = ["postgresql"], version = "^3.17.7"}
rich = "^13.9.3"
psycopg2-binary = "^2.9.10"
tqdm = "^4.66.5"
strawberry-graphql = "^0.247.0"
gql = "^3.5.0"
cachetools = "^5.5.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
